requirements.txt
# CryptoSignal Hub - Dependencies
# 交易所数据统一接口
ccxt>=4.4.0
# 美股/宏观数据
yfinance>=0.2.40
# 数据处理
pandas>=2.2.0
# 技术指标（纯 Python，无需编译 TA-Lib）
ta>=0.11.0
# HTTP 请求
requests>=2.31.0
httpx>=0.27.0
# 定时任务
apscheduler>=3.10.4
# Web 框架
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
# AI 分析（兼容 DeepSeek / OpenAI）
openai>=1.40.0
# 配置管理
pyyaml>=6.0.1
pydantic>=2.8.0
pydantic-settings>=2.4.0
# 时区处理
pytz>=2024.1
